# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: File Sync WebApp
  description: Web Frontend for File Sync Service
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: ''
  version: '0.12.4'
paths:
...
  /sync/sync_calendar:
    post:
      operationId: sync_calendar
      responses:
        '200':
          description: Sync Calendar DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_garmin:
    post:
      operationId: sync_garmin
      responses:
        '200':
          description: Sync Garmin DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_movie:
    post:
      operationId: sync_movie
      responses:
        '200':
          description: Sync Movie DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /sync/sync_security:
    post:
      operationId: sync_security
      responses:
        '200':
          description: Sync Security Logs
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_weather:
    post:
      operationId: sync_weather
      responses:
        '200':
          description: Sync Weather Data
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
components:
  schemas:
...
    str:
      type: string